Component,ComponentName,ComponentList,GameObjcet,Sceneをユーザとエンジンで共有するフォルダに移動
