+++ Engine/Core/Core.vcxproj.filters
@@ -22,15 +22,11 @@
     <ClCompile Include="src\yougine\Editor\SelectionInfo.cpp" />
     <ClCompile Include="src\yougine\managers\GameManager.cpp" />
     <ClCompile Include="src\yougine\managers\RenderManager.cpp" />
-    <ClCompile Include="src\yougine\components\Component.cpp" />
     <ClCompile Include="src\yougine\Editor\EditorWindow.cpp" />
     <ClCompile Include="src\yougine\Editor\EditorWindowsManager.cpp" />
     <ClCompile Include="src\yougine\Editor\HierarchyWindow.cpp" />
     <ClCompile Include="src\yougine\InputManager.cpp" />
-    <ClCompile Include="src\yougine\managers\ComponentList.cpp" />
     <ClCompile Include="src\yougine\managers\CustomScriptManager.cpp" />
-    <ClCompile Include="src\yougine\GameObject.cpp" />
-    <ClCompile Include="src\yougine\Scene.cpp" />
     <ClCompile Include="src\yougine\components\TransformComponent.cpp" />
     <ClCompile Include="src\yougine\SceneFiles\SceneFileExporter.cpp" />
     <ClCompile Include="src\yougine\SceneExport_tamesiMain.cpp" />
@@ -186,16 +182,11 @@
     <ClInclude Include="src\yougine\managers\GameManager.h" />
     <ClInclude Include="src\yougine\managers\IManager.h" />
     <ClInclude Include="src\yougine\managers\RenderManager.h" />
-    <ClInclude Include="src\yougine\components\Component.h" />
     <ClInclude Include="src\yougine\Editor\EditorWindow.h" />
     <ClInclude Include="src\yougine\Editor\EditorWindowsManager.h" />
     <ClInclude Include="src\yougine\Editor\HierarchyWindow.h" />
     <ClInclude Include="src\yougine\InputManager.h" />
-    <ClInclude Include="src\yougine\managers\ComponentList.h" />
-    <ClInclude Include="src\yougine\components\ComponentName.h" />
     <ClInclude Include="src\yougine\managers\CustomScriptManager.h" />
-    <ClInclude Include="src\yougine\GameObject.h" />
-    <ClInclude Include="src\yougine\Scene.h" />
     <ClInclude Include="src\yougine\components\TransformComponent.h" />
     <ClInclude Include="src\yougine\SceneFiles\SceneFileExporter.h" />
     <ClInclude Include="src\yougine\SceneFiles\SceneLoader.h" />
