+++ Engine/Core/src/yougine/Editor/ComponentsIncluder.h
@@ -1,5 +1,5 @@
 // #pragma once
-#include "../components/Component.h"
+#include "UserShare/components/Component.h"
 #include "UserShare/utilitys/YougineMath.h"
 //#include "../components/RenderComponent.h"
 //#include "../components/TransformComponent.h"
