+++ Engine/Core/src/yougine/SceneFiles/SceneFileExporter.cpp
@@ -5,7 +5,7 @@
 #include <typeinfo>
 #include <iostream>
 
-#include "../components/Component.h"
+#include "UserShare/components/Component.h"
 #include "UserShare/utilitys/Split.h"
 #include "UserShare/utilitys/YougineMath.h"
