+++ Engine/Core/src/yougine/components/Camera/CameraComponent.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 #include <memory>
 
-#include "../Component.h"
+#include "UserShare/components/Component.h"
 #include "../TransformComponent.h"
 
 namespace yougine::components::camera
@@ -9,7 +9,7 @@ namespace yougine::components::camera
     class CameraComponent : public yougine::components::Component
     {
     private:
-         static std::shared_ptr<CameraComponent> main_camera;
+        static std::shared_ptr<CameraComponent> main_camera;
 
     public:
         CameraComponent();
