+++ Engine/Core/src/yougine/managers/ComponentExportParameterManager.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <vector>
 
-#include "../components/Component.h"
+#include "UserShare/components/Component.h"
 #include "UserShare/utilitys/uuid/YougineUuid.h"
 #include "./../components/ComponentExportParameters/ComponentExportParameter.h"
 #include "./../Editor/ProjectWindows/Assets/element/Model/Asset.h"
