+++ Engine/UserEngineCommon/UserEngineCommon.vcxproj
@@ -19,14 +19,23 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="include\UserShare\components\Component.h" />
+    <ClInclude Include="include\UserShare\components\ComponentName.h" />
+    <ClInclude Include="include\UserShare\GameObject.h" />
     <ClInclude Include="include\UserShare\Layer.h" />
     <ClInclude Include="include\UserShare\MacroDifHeader.h" />
+    <ClInclude Include="include\UserShare\managers\ComponentList.h" />
+    <ClInclude Include="include\UserShare\Scene.h" />
     <ClInclude Include="include\UserShare\UserScriptComponent.h" />
     <ClInclude Include="include\UserShare\utilitys\Split.h" />
     <ClInclude Include="include\UserShare\utilitys\uuid\YougineUuid.h" />
     <ClInclude Include="include\UserShare\utilitys\YougineMath.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="include\UserShare\components\Component.cpp" />
+    <ClCompile Include="include\UserShare\GameObject.cpp" />
+    <ClCompile Include="include\UserShare\managers\ComponentList.cpp" />
+    <ClCompile Include="include\UserShare\Scene.cpp" />
     <ClCompile Include="include\UserShare\UserScriptComponent.cpp" />
     <ClCompile Include="include\UserShare\utilitys\Split.cpp" />
     <ClCompile Include="include\UserShare\Layer.cpp" />
